+++ qsprpred/data/tests.py
@@ -35,7 +35,7 @@ from .utils.descriptorcalculator import (
 )
 from .utils.descriptorsets import (
     DataFrameDescriptorSet, DescriptorSet, DrugExPhyschem, FingerprintSet,
-    PredictorDesc, RDKitDescs, TanimotoDistances
+    PredictorDesc, RDKitDescs, TanimotoDistances, SmilesDesc
 )
 from .utils.feature_standardization import SKLearnStandardizer
 from .utils.featurefilters import BorutaFilter, HighCorrelationFilter, LowVarianceFilter
@@ -1401,6 +1401,14 @@ class TestDescriptorSets(DataSetsMixIn, TestCase):
         self.assertEqual(
             self.dataset.X.shape, (len(self.dataset), len(Descriptors._descList) + 10)
         )
+
+    def testSmilesDesc(self):
+        """Test the smiles descriptors calculator."""
+        desc_calc = MoleculeDescriptorsCalculator([SmilesDesc()])
+        self.dataset.addDescriptors(desc_calc)
+
+        self.assertEqual(self.dataset.X.shape, (len(self.dataset), 1))
+        self.assertTrue(self.dataset.X.any().any())
 
     def testConsistency(self):
         """Test if the descriptor calculator is consistent with the dataset."""