+++ CHANGELOG.md
@@ -25,6 +25,7 @@ From v2.0.1 to v2.1.0.dev0
 - A new tutorial `adding_new_components.ipynb` has been added to the `tutorials` folder, which demonstrates how to add new model to QSPRpred.
 - A new function `Metrics.checkMetricCompatibility` has been added, which checks if a metric is compatible with a given task and a given prediction methods (i.e. `predict` or `predictProba`)
 - In `EvaluationMethod` (see above), an attribute `use_proba` has been added, which determines whether the `predict` or `predictProba` method is used to make predictions (#56).
+- Add new descriptorset `SmilesDesc` to use the smiles strings as a descriptor.
 - New module `early_stopping` with classes `EarlyStopping` and `EarlyStoppingMode` has been added. This module allows for more control over early stopping in models that support it.
 - Add new descriptorset `SmilesDesc` to use the smiles strings as a descriptor.
 - Refactoring of the test suite under `qsprpred.data` and improvement of temporary file handling (!114).